NASDAQ Slash Command for Mixmax
This slash command works by calling nasdaq.com's symbol search and for
each symbol querying google finance's quote. Once a symbol is selected,
the resolver formats it to color the numbers that are part of the
information to red if it performed poorly during the last day or green
otherwise.